Extract `useWorkspaces`, filters logic
